backend/requirements.txt
# FastAPI and server dependencies
fastapi>=0.105.0
uvicorn>=0.24.0
python-multipart>=0.0.6
aiofiles>=23.2.1
pydantic>=2.5.2
pydantic_settings>=2.1.0
sqlalchemy>=2.0.23
alembic>=1.12.1
python-jose>=3.3.0
passlib>=1.7.4

# Document processing
PyPDF2>=3.0.1
python-docx>=1.0.1
pandas>=2.1.3
beautifulsoup4>=4.12.2
markdown>=3.5.1
python-magic>=0.4.27

# Vector database and embeddings
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4  # Using FAISS instead of HNSWLIB

# LLM integrations
openai>=1.3.7
anthropic>=0.7.7
google-generativeai>=0.3.2
aiohttp>=3.8.3  # Required for OpenRouter API calls

# Utilities
python-dotenv>=1.0.0
tenacity>=8.2.3
loguru>=0.7.2